pyproject.toml
[tool.poetry]
name = "social-media-generator"
version = "0.1.0"
description = "AI-powered social media content generator"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.9"
torch = { version = "2.2.1", extras = ["cpu"] }
transformers = ">=4.38.2"
fastapi = "^0.109.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.34.0"
tweepy = "^4.14.0"
numpy = "<2.0.0"
"Mastodon.py" = "^1.8.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
ruff = "^0.2.2"
uv = "^0.1.13"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"